refactor: add dataloader types file
