+++ src/dataloader/dataloader.service.ts
@@ -21,37 +21,7 @@ import { User } from "../users/models/user.model";
 import { UsersService } from "../users/users.service";
 import { Vote } from "../votes/models/vote.model";
 import { VotesService } from "../votes/votes.service";
-
-export interface IsLikedByMeKey {
-  userId: number;
-  postId: number;
-}
-
-export interface Dataloaders {
-  // Proposals & Votes
-  proposalActionsLoader: DataLoader<number, ProposalAction>;
-  proposalImagesLoader: DataLoader<number, Image[]>;
-  proposalVoteCountLoader: DataLoader<number, number>;
-  proposalVotesLoader: DataLoader<number, Vote[]>;
-
-  // Posts
-  isPostLikedByMeLoader: DataLoader<IsLikedByMeKey, boolean>;
-  postImagesLoader: DataLoader<number, Image[]>;
-  postLikeCountLoader: DataLoader<number, number>;
-  postLikesLoader: DataLoader<number, Like[]>;
-
-  // Groups
-  groupCoverPhotosLoader: DataLoader<number, Image>;
-  groupMemberCountLoader: DataLoader<number, number>;
-  groupMembersLoader: DataLoader<number, GroupMember[]>;
-  groupsLoader: DataLoader<number, Group>;
-  memberRequestCountLoader: DataLoader<number, number>;
-
-  // Misc.
-  profilePicturesLoader: DataLoader<number, Image>;
-  roleMemberCountLoader: DataLoader<number, number>;
-  usersLoader: DataLoader<number, User>;
-}
+import { Dataloaders, IsLikedByMeKey } from "./dataloader.types";
 
 @Injectable()
 export class DataloaderService {
