+++ src/posts/posts.service.ts
@@ -2,16 +2,16 @@ import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { FileUpload } from "graphql-upload";
 import { FindOptionsWhere, In, Repository } from "typeorm";
+import { IsLikedByMeKey } from "../dataloader/dataloader.types";
 import { deleteImageFile, saveImage } from "../images/image.utils";
 import { ImagesService } from "../images/images.service";
 import { Image } from "../images/models/image.model";
 import { LikesService } from "../likes/likes.service";
+import { Like } from "../likes/models/like.model";
 import { User } from "../users/models/user.model";
 import { CreatePostInput } from "./models/create-post.input";
 import { Post } from "./models/post.model";
-import { Like } from "../likes/models/like.model";
 import { UpdatePostInput } from "./models/update-post.input";
-import { IsLikedByMeKey } from "../dataloader/dataloader.service";
 
 type PostWithLikeCount = Post & { likeCount: number };
