+++ src/server-invites/server-invites.resolver.ts
@@ -9,7 +9,7 @@ import {
   Resolver,
 } from "@nestjs/graphql";
 import { CurrentUser } from "../auth/decorators/current-user.decorator";
-import { Dataloaders } from "../dataloader/dataloader.service";
+import { Dataloaders } from "../dataloader/dataloader.types";
 import { User } from "../users/models/user.model";
 import { CreateServerInviteInput } from "./models/create-server-invite.input";
 import { CreateServerInvitePayload } from "./models/create-server-invite.payload";
